Fix delay when bookmarking an item on the initial load
During the initial load, the view is not observing the DB yet. The view only receives the initial data (the current data in the DB) and a loading status. This allows the user to view the current data while new data is being fetched from the API. However, if the user bookmark (or delete a bookmark), the icon is not updated right away. The icon is only updated when the new values have fetched and sent to the UI. As a result, even though the action has been registered correctly in the DB, the user doesn't see the icon change until the next update. We simply fix this by updating the icon right away in `RecallAdapter` instead of waiting for the next update.